Restructure and install subfolders in /components
