+++ app/components/Common/Footer.js
@@ -1,4 +1,3 @@
-// components/Footer.js
 export default function Footer() {
   return (
     <footer className='bg-gray-800 text-white fixed bottom-0 left-0 w-full p-4 shadow-md'>
